refactor cadet_dll into multiple files, simulations run now
