+++ cadet/cadet_dll.py
@@ -2,9 +2,7 @@
 import ctypes
 import numpy
 import addict
-import time
-
-#input("Press Enter to continue...")
+import cadet.cadet_dll_parameterprovider as cadet_dll_parameterprovider
 
 def log_handler(file, func, line, level, level_name, message):
     log_print('{} ({}:{:d}) {}'.format(level_name.decode('utf-8') , func.decode('utf-8') , line, message.decode('utf-8') ))
@@ -23,34 +21,6 @@ if 0:
     log_print = print
 else:
     log_print = null
-
-
-class PARAMETERPROVIDER(ctypes.Structure):
-
-    _fields_ = [
-        ('userData', ctypes.py_object),
-
-        ('getDouble', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.POINTER(ctypes.c_double))),
-        ('getInt', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, point_int)),
-        ('getBool', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.POINTER(ctypes.c_uint8))),
-        ('getString', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.POINTER(ctypes.c_char_p))),
-
-        ('getDoubleArray', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, point_int, array_double)),
-        ('getIntArray', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, point_int, ctypes.POINTER(point_int))),
-        ('getBoolArray', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, point_int, ctypes.POINTER(ctypes.POINTER(ctypes.c_uint8)))),
-        ('getStringArray', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, point_int, ctypes.POINTER(ctypes.POINTER(ctypes.c_char_p)))),
-
-        ('getDoubleArrayItem', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.c_int, ctypes.POINTER(ctypes.c_double))),
-        ('getIntArrayItem', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.c_int, point_int)),
-        ('getBoolArrayItem', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.c_int, ctypes.POINTER(ctypes.c_uint8))),
-        ('getStringArrayItem', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.c_int, ctypes.POINTER(ctypes.c_char_p))),
-
-        ('exists', ctypes.CFUNCTYPE(ctypes.c_int, ctypes.py_object, ctypes.c_char_p)),
-        ('isArray', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p, ctypes.POINTER(ctypes.c_uint8))),
-        ('numElements', ctypes.CFUNCTYPE(ctypes.c_int, ctypes.py_object, ctypes.c_char_p)),
-        ('pushScope', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object, ctypes.c_char_p)),
-        ('popScope', ctypes.CFUNCTYPE(c_cadet_result, ctypes.py_object)),
-    ]
 
 
 class CADETAPIV010000_DATA():
@@ -105,7 +75,7 @@ class CADETAPIV010000_DATA():
         'nComp': point_int,
         'nBound': point_int,
         None: None,
-        'parameterProvider': ctypes.POINTER(PARAMETERPROVIDER)
+        'parameterProvider': ctypes.POINTER(cadet_dll_parameterprovider.PARAMETERPROVIDER)
     }
 
     lookup_call = {
@@ -135,296 +105,6 @@ def setup_api():
 class CADETAPIV010000(ctypes.Structure):
     _fields_ = setup_api()
 
-
-class NestedDictReader:
-
-    def __init__(self, data):
-        self._root = data
-        self._cursor = []
-        self._cursor.append(data)
-        self.buffer = None
-
-    def push_scope(self, scope):
-        if scope in self._cursor[-1]:
-            log_print('Entering scope {}'.format(scope))
-            self._cursor.append(self._cursor[-1][scope])
-            return True
-
-        return False
-
-    def pop_scope(self):
-        self._cursor.pop()
-        log_print('Exiting scope')
-
-    def current(self):
-        return self._cursor[-1]
-
-
-def param_provider_get_double(reader, name, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-        try:
-            float_val = float(o)
-        except TypeError:
-            float_val = float(o[0])
-
-        val[0] = ctypes.c_double(float_val)
-
-        log_print('GET scalar [double] {}: {}'.format(n, float(val[0])))
-        return 0
-
-    return -1
-
-
-def param_provider_get_int(reader, name, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-        try:
-            int_val = int(o)
-        except TypeError:
-            int_val = int(o[0])
-
-        val[0] = ctypes.c_int(int_val)
-
-        log_print('GET scalar [int] {}: {}'.format(n, int(val[0])))
-        return 0
-
-    return -1
-
-
-def param_provider_get_bool(reader, name, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-
-        try:
-            int_val = int(o)
-        except TypeError:
-            int_val = int(o[0])
-
-        val[0] = ctypes.c_uint8(int_val)
-
-        log_print('GET scalar [bool] {}: {}'.format(n, bool(val[0])))
-        return 0
-
-    return -1
-
-
-def param_provider_get_string(reader, name, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-
-        #we have one of array of strings, array of bytestrings, bytestring or string or something convertable to one of these
-        if hasattr(o, 'encode'):
-            bytes_val = o.encode('utf-8')
-        elif hasattr(o, 'decode'):
-            bytes_val = o
-        elif hasattr(o[0], 'encode'):
-            bytes_val = o[0].encode('utf-8')
-        elif hasattr(o[0], 'decode'):
-            bytes_val = o[0]
-
-        reader.buffer = bytes_val
-        val[0] = ctypes.cast(reader.buffer, ctypes.c_char_p)
-
-        return 0
-
-    return -1
-
-
-def param_provider_get_double_array(reader, name, n_elem, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if (n in c):
-        o = c[n]
-        if isinstance(o, list):
-            o = numpy.ascontiguousarray(o)
-        if (not isinstance(o, numpy.ndarray)) or (o.dtype != numpy.double) or (not o.flags.c_contiguous):
-            return -1
-
-        n_elem[0] = ctypes.c_int(o.size)
-        val[0] = o.ctypes.data_as(ctypes.POINTER(ctypes.c_double))
-        log_print('GET array [double] {}: {}'.format(n, o))
-        return 0
-
-    return -1
-
-
-def param_provider_get_int_array(reader, name, n_elem, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if (n in c):
-        o = c[n]
-        if isinstance(o, list):
-            o = numpy.ascontiguousarray(o)
-        if (not isinstance(o, numpy.ndarray)) or (o.dtype != numpy.int) or (not o.flags.c_contiguous):
-            return -1
-
-        n_elem[0] = ctypes.c_int(o.size)
-        val[0] = o.ctypes.data_as(ctypes.POINTER(ctypes.c_int))
-        log_print('GET array [int] {}: {}'.format(n, o))
-        return 0
-
-    return -1
-
-
-def param_provider_get_double_array_item(reader, name, index, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-
-        try:
-            float_val = float(o)
-        except TypeError:
-            float_val = float(o[index])
-
-        val[0] = ctypes.c_double(float_val)
-        log_print('GET array [double] ({}) {}: {}'.format(index, n, val[0]))
-        return 0
-
-    return -1
-
-
-def param_provider_get_int_array_item(reader, name, index, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-
-        try:
-            int_val = int(o)
-        except TypeError:
-            int_val = int(o[index])
-
-        val[0] = ctypes.c_int(int_val)
-        log_print('GET array [int] ({}) {}: {}'.format(index, n, val[0]))
-        return 0
-
-    return -1
-
-
-def param_provider_get_bool_array_item(reader, name, index, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-
-        try:
-            int_val = int(o)
-        except TypeError:
-            int_val = int(o[index])
-
-        val[0] = ctypes.c_uint8(int_val)
-        log_print('GET array [bool] ({}) {}: {}'.format(index, n, bool(val[0])))
-        return 0
-
-    return -1
-
-
-def param_provider_get_string_array_item(reader, name, index, val):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n in c:
-        o = c[n]
-        if index == 0:
-            if hasattr(o, 'encode'):
-                bytes_val = o.encode('utf-8')
-            elif hasattr(o, 'decode'):
-                bytes_val = o
-        else:
-            if hasattr(o[index], 'encode'):
-                bytes_val = o[index].encode('utf-8')
-            elif hasattr(o[index], 'decode'):
-                bytes_val = o[index]
-
-        reader.buffer = bytes_val
-        val[0] = ctypes.cast(reader.buffer, ctypes.c_char_p)
-        log_print('GET array [string] ({}) {}: {}'.format(index, n, reader.buffer.decode('utf-8')))
-        return 0
-
-    return -1
-
-
-def param_provider_exists(reader, name):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    log_print('EXISTS {}: {}'.format(n, n in c))
-
-    if n in c:
-        return 1
-
-    return 0
-
-
-def param_provider_is_array(reader, name, res):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n not in c:
-        return -1
-
-    o = c[n]
-    res[0] = ctypes.c_uint8(0)
-    if isinstance(o, list):
-        res[0] = ctypes.c_uint8(1)
-    elif isinstance(o, numpy.ndarray):
-        res[0] = ctypes.c_uint8(1)
-
-    log_print('ISARRAY {}: {}'.format(n, bool(res[0])))
-
-    return 0
-
-
-def param_provider_num_elements(reader, name):
-    n = name.decode('utf-8')
-    c = reader.current()
-
-    if n not in c:
-        return -1
-
-    o = c[n]
-    if isinstance(o, list):
-        log_print('NUMELEMENTS {}: {}'.format(n, len(o)))
-        return len(o)
-    elif isinstance(o, numpy.ndarray):
-        log_print('NUMELEMENTS {}: {}'.format(n, o.size))
-        return o.size
-
-    log_print('NUMELEMENTS {}: {}'.format(n, 1))
-    return 1
-
-
-def param_provider_push_scope(reader, name):
-    n = name.decode('utf-8')
-
-    if reader.push_scope(n):
-        return 0
-    else:
-        return -1
-
-
-def param_provider_pop_scope(reader):
-    reader.pop_scope()
-    return 0
 
 def null(obj):
     "do nothing"
@@ -625,32 +305,7 @@ class CadetDLL:
 
 
     def run(self, filename = None, simulation=None, timeout = None, check=None):
-        pp = PARAMETERPROVIDER()
-
-        sim_input = recursively_convert_dict(simulation)
-
-        pp.userData = NestedDictReader(sim_input)
-
-        pp.getDouble = PARAMETERPROVIDER._fields_[1][1](param_provider_get_double)
-        pp.getInt = PARAMETERPROVIDER._fields_[2][1](param_provider_get_int)
-        pp.getBool = PARAMETERPROVIDER._fields_[3][1](param_provider_get_bool)
-        pp.getString = PARAMETERPROVIDER._fields_[4][1](param_provider_get_string)
-
-        pp.getDoubleArray = PARAMETERPROVIDER._fields_[5][1](param_provider_get_double_array)
-        pp.getIntArray = PARAMETERPROVIDER._fields_[6][1](param_provider_get_int_array)
-        pp.getBoolArray = ctypes.cast(None, PARAMETERPROVIDER._fields_[7][1])
-        pp.getStringArray = ctypes.cast(None, PARAMETERPROVIDER._fields_[8][1])
-
-        pp.getDoubleArrayItem = PARAMETERPROVIDER._fields_[9][1](param_provider_get_double_array_item)
-        pp.getIntArrayItem = PARAMETERPROVIDER._fields_[10][1](param_provider_get_int_array_item)
-        pp.getBoolArrayItem = PARAMETERPROVIDER._fields_[11][1](param_provider_get_bool_array_item)
-        pp.getStringArrayItem = PARAMETERPROVIDER._fields_[12][1](param_provider_get_string_array_item)
-
-        pp.exists = PARAMETERPROVIDER._fields_[13][1](param_provider_exists)
-        pp.isArray = PARAMETERPROVIDER._fields_[14][1](param_provider_is_array)
-        pp.numElements = PARAMETERPROVIDER._fields_[15][1](param_provider_num_elements)
-        pp.pushScope = PARAMETERPROVIDER._fields_[16][1](param_provider_push_scope)
-        pp.popScope = PARAMETERPROVIDER._fields_[17][1](param_provider_pop_scope)
+        pp = cadet_dll_parameterprovider.PARAMETERPROVIDER(simulation)
 
         self._api.runSimulation(self._driver, ctypes.byref(pp))
         self.res = SimulationResult(self._api, self._driver)
@@ -678,14 +333,14 @@ class CadetDLL:
                         solution[key][solution_str] = out
         return solution
 
-    def load_solution_io(self, sim, solution, solution_str):
+    def load_solution_io(self, sim, solution_fun, solution_str):
         solution = addict.Dict()
         if self.res is not None:
             for key,value in sim.root.input['return'].items():
                 if key.startswith('unit'):
                     if value[f'write_{solution_str}']:
                         unit = int(key[-3:])
-                        t, out = self.res.inlet(unit)
+                        t, out = solution_fun(unit)
 
                         if not len(solution.solution_times):
                             solution.solution_times = t
@@ -720,67 +375,67 @@ class CadetDLL:
         return self.load_solution(sim, self.res.volume, 'solution_volume')
     
     def load_derivative_inlet(self, sim):
-        return self.load_solution_io(sim, self.res.soldot_inlet, 'soldot_inlet')
+        return self.load_solution_io(sim, self.res.derivativeInlet, 'soldot_inlet')
     
     def load_derivative_outlet(self, sim):
-        return self.load_solution_io(sim, self.res.soldot_outlet, 'soldot_outlet')
+        return self.load_solution_io(sim, self.res.derivativeOutlet, 'soldot_outlet')
     
     def load_derivative_bulk(self, sim):
-        return self.load_solution(sim, self.res.soldot_bulk, 'soldot_bulk')
+        return self.load_solution(sim, self.res.derivativeBulk, 'soldot_bulk')
 
     def load_derivative_particle(self, sim):
-        return self.load_solution(sim, self.res.soldot_particle, 'soldot_particle')
+        return self.load_solution(sim, self.res.derivativeParticle, 'soldot_particle')
 
     def load_derivative_solid(self, sim):
-        return self.load_solution(sim, self.res.soldot_solid, 'soldot_solid')
+        return self.load_solution(sim, self.res.derivativeSolid, 'soldot_solid')
 
     def load_derivative_flux(self, sim):
-        return self.load_solution(sim, self.res.soldot_flux, 'soldot_flux')
+        return self.load_solution(sim, self.res.derivativeFlux, 'soldot_flux')
 
     def load_derivative_volume(self, sim):
-        return self.load_solution(sim, self.res.soldot_volume, 'soldot_volume')
+        return self.load_solution(sim, self.res.derivativeVolume, 'soldot_volume')
 
     def load_sensitivity_inlet(self, sim):
-        return self.load_solution_io(sim, self.res.sens_inlet, 'sens_inlet')
+        return self.load_solution_io(sim, self.res.sensitivityInlet, 'sens_inlet')
 
     def load_sensitivity_outlet(self, sim):
-        return self.load_solution_io(sim, self.res.sens_outlet, 'sens_outlet')
+        return self.load_solution_io(sim, self.res.sensitivityOutlet, 'sens_outlet')
 
     def load_sensitivity_bulk(self, sim):
-        return self.load_solution(sim, self.res.sens_bulk, 'sens_bulk')
+        return self.load_solution(sim, self.res.sensitivityBulk, 'sens_bulk')
 
     def load_sensitivity_particle(self, sim):
-        return self.load_solution(sim, self.res.sens_particle, 'sens_particle')
+        return self.load_solution(sim, self.res.sensitivityParticle, 'sens_particle')
 
     def load_sensitivity_solid(self, sim):
-        return self.load_solution(sim, self.res.sens_solid, 'sens_solid')
+        return self.load_solution(sim, self.res.sensitivitySolid, 'sens_solid')
 
     def load_sensitivity_flux(self, sim):
-        return self.load_solution(sim, self.res.sens_flux, 'sens_flux')
+        return self.load_solution(sim, self.res.sensitivityFlux, 'sens_flux')
 
     def load_sensitivity_volume(self, sim):
-        return self.load_solution(sim, self.res.sens_volume, 'sens_volume')
+        return self.load_solution(sim, self.res.sensitivityVolume, 'sens_volume')
 
     def load_sensitivity_derivative_inlet(self, sim):
-        return self.load_solution_io(sim, self.res.sensdot_inlet, 'sensdot_inlet')
+        return self.load_solution_io(sim, self.res.sensitivityDerivativeInlet, 'sensdot_inlet')
 
     def load_sensitivity_derivative_outlet(self, sim):
-        return self.load_solution_io(sim, self.res.sensdot_outlet, 'sensdot_outlet')
+        return self.load_solution_io(sim, self.res.sensitivityDerivativeOutlet, 'sensdot_outlet')
 
     def load_sensitivity_derivative_bulk(self, sim):
-        return self.load_solution(sim, self.res.sensdot_bulk, 'sensdot_bulk')
+        return self.load_solution(sim, self.res.sensitivityDerivativeBulk, 'sensdot_bulk')
 
     def load_sensitivity_derivative_particle(self, sim):
-        return self.load_solution(sim, self.res.sensdot_particle, 'sensdot_particle')
+        return self.load_solution(sim, self.res.sensitivityDerivativeParticle, 'sensdot_particle')
 
     def load_sensitivity_derivative_solid(self, sim):
-        return self.load_solution(sim, self.res.sensdot_solid, 'sensdot_solid')
+        return self.load_solution(sim, self.res.sensitivityDerivativeSolid, 'sensdot_solid')
 
     def load_sensitivity_derivative_flux(self, sim):
-        return self.load_solution(sim, self.res.sensdot_flux, 'sensdot_flux')
+        return self.load_solution(sim, self.res.sensitivityDerivativeFlux, 'sensdot_flux')
 
     def load_sensitivity_derivative_volume(self, sim):
-        return self.load_solution(sim, self.res.sensdot_volume, 'sensdot_volume')
+        return self.load_solution(sim, self.res.sensitivityDerivativeVolume, 'sensdot_volume')
 
     def load_results(self, sim):
         sim.root.output.solution.update(self.load_inlet(sim))
@@ -797,28 +452,17 @@ class CadetDLL:
         sim.root.output.solution.update(self.load_derivative_solid(sim))
         sim.root.output.solution.update(self.load_derivative_flux(sim))
         sim.root.output.solution.update(self.load_derivative_volume(sim))
-        sim.root.output.solution.update(self.load_sensitivity_inlet(sim))
-        sim.root.output.solution.update(self.load_sensitivity_outlet(sim))
-        sim.root.output.solution.update(self.load_sensitivity_bulk(sim))
-        sim.root.output.solution.update(self.load_sensitivity_particle(sim))
-        sim.root.output.solution.update(self.load_sensitivity_solid(sim))
-        sim.root.output.solution.update(self.load_sensitivity_flux(sim))
-        sim.root.output.solution.update(self.load_sensitivity_volume(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_inlet(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_outlet(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_bulk(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_particle(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_solid(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_flux(sim))
-        sim.root.output.solution.update(self.load_sensitivity_derivative_volume(sim))
-
-
-def recursively_convert_dict(data):
-    ans = addict.Dict()
-    for key_original, item in data.items():
-        if isinstance(item, dict):
-            ans[key_original] = recursively_convert_dict(item)
-        else:
-            key = str.upper(key_original)
-            ans[key] = item
-    return ans
+        #sim.root.output.solution.update(self.load_sensitivity_inlet(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_outlet(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_bulk(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_particle(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_solid(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_flux(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_volume(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_inlet(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_outlet(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_bulk(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_particle(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_solid(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_flux(sim))
+        #sim.root.output.solution.update(self.load_sensitivity_derivative_volume(sim))
